src/main/webapp/app/entities/configs/delete/configs-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import {ComponentFixture, fakeAsync, inject, TestBed, tick} from '@angular/core/testing';
import {HttpResponse} from '@angular/common/http';
import {HttpClientTestingModule} from '@angular/common/http/testing';
import {of} from 'rxjs';
import {NgbActiveModal} from '@ng-bootstrap/ng-bootstrap';

import {ConfigsService} from '../service/configs.service';

import {ConfigsDeleteDialogComponent} from './configs-delete-dialog.component';

describe('Configs Management Delete Component', () => {
  let comp: ConfigsDeleteDialogComponent;
  let fixture: ComponentFixture<ConfigsDeleteDialogComponent>;
  let service: ConfigsService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [ConfigsDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(ConfigsDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(ConfigsDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(ConfigsService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      })
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
